+++ include/EVT/io/platform/f3xx/PWMf3xx.hpp
@@ -18,7 +18,7 @@ public:
      */
     PWMf3xx(Pin pin);
 
-    void setDutyCycle(float dutyCycle);
+    void setDutyCycle(uint32_t dutyCycle);
 
     void setPeriod(uint32_t period);
